+++ src/muos/steps/select_pacman_mirrors.py
@@ -1,9 +1,9 @@
 from ..environment import Environment
 from ..get_lines_as_list import get_lines_as_list
 from ..get_static_dir import get_static_dir
-from ..select_step import SelectStep
+from ..multi_select_step import MultiSelectStep
 
-class SelectPacmanMirrors(SelectStep):
+class SelectPacmanMirrors(MultiSelectStep):
     name: str = 'Select Pacman mirrors'
 
     def main(self, environment: Environment) -> None: